inode now appears in the userspace event

+++ pkg/security/ebpf/c/include/constants/offsets/filesystem.h
@@ -139,9 +139,9 @@ int __attribute__((always_inline)) get_mount_mount_id(void *mnt) {
     return mount_id;
 }
 
-u64 __attribute__((always_inline)) get_mount_mount_ns_inum(void *mnt) {
+u32 __attribute__((always_inline)) get_mount_mount_ns_inum(void *mnt) {
     void* mnt_ns = NULL;
-    u64   inum = 0;
+    u32   inum = 0;
 
     bpf_probe_read(&mnt_ns, sizeof(mnt_ns), mnt + get_mount_offset_of_mount_ns());
     if (!mnt_ns) {
